Move clear_* functions to utils.
