+++ utils.h
@@ -1,0 +1,14 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+/**
+ * clear_update_buffer() - Clear the update buffer
+ */
+void clear_update_buffer(void);
+
+/**
+ * clear_stamps() - Clear the on screen stamp buffer
+ */
+void clear_stamps(void);
+
+#endif
